api-service/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.0
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pymongo==4.6.1
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
